push the fourth view
