+++ Usport/Modules/Leagues/view/TableViewController.swift
@@ -25,6 +25,10 @@ class TableViewController: UITableViewController {
             }
         }
     }
+
+    override func tableView(_ tableView: UITableView, heightForRowAt indexPath: IndexPath) -> CGFloat {
+        return 150
+    }
     override func tableView(_ tableView: UITableView, numberOfRowsInSection section: Int) -> Int {
         return (viewModel?.leagues.count) ?? 0
     }
@@ -39,13 +43,11 @@ class TableViewController: UITableViewController {
         return cell
     }
     
-    override func tableView(_ tableView: UITableView, heightForRowAt indexPath: IndexPath) -> CGFloat {
-        return 150
-    }
+   
     override func tableView(_ tableView: UITableView, didSelectRowAt indexPath: IndexPath) {
         tableView.deselectRow(at: indexPath, animated: true)
         
-        let vc = storyboard?.instantiateViewController(withIdentifier: "LVC") as! leaguesDetails
+        let vc = storyboard?.instantiateViewController(withIdentifier: "LVC") as! LeaguesDetailsViewController
         let selectedLeague = viewModel?.leagues[indexPath.row]
         
         let leaguesDetailsViewModel = LeaguesDetailsViewModel()
